a bunch of changes
- created 3 boss sprites, and 3 boss bullets
- created 1 boss prefab, attached EnemyBoss.cs to it
- boss faces player and shoots regular bullet
- deleted EnemyCommon.cs
- attached Enemy.cs to all common enemies instead
- changed health Slider to an Image
- changed all circle colliders to polygon colliders to fit the outline of the ship better

+++ Skyfall 2.0/Skyfall 2.0/Assets/_Scripts/Characters/Players/Player.cs
@@ -25,8 +25,8 @@ public class Player : ScriptableObject
     public float maxHealth;
     public float attack;
     public float defense;
-    public int moveSpeed;
-    public int shotDensity;
+    public float moveSpeed;
+    public float shotDensity;
     public float specialCooldown;       //how long the player has to wait before using another special
 
     public const int maxLives = 3;
